requirements.txt
streamlit==1.55.0
psycopg2-binary==2.9.11
requests==2.32.5
youtube-transcript-api==1.2.4
pandas==2.2.3